test(pagination): recategorize and clean up browser tests
